+++ tests/admin_utils.cpp
@@ -20,7 +20,8 @@
 #include <sstream>
 #include <thread>
 #include <iostream>
-#include <cpprealm/internal/generic_network_transport.hpp>
+#include <cpprealm/networking/http.hpp>
+#include <cpprealm/internal/networking/utils.hpp>
 
 #include "admin_utils.hpp"
 #include "external/json/json.hpp"
@@ -30,12 +31,12 @@ namespace Admin {
     std::mutex Admin::Session::mutex;
 
     static app::Response do_http_request(app::Request &&request) {
-        internal::DefaultTransport transport;
+        networking::default_http_transport transport;
         std::promise<app::Response> p;
         std::future<app::Response> f = p.get_future();
-        transport.send_request_to_server(std::move(request),
-                                         [&p](auto &&response) {
-                                             p.set_value(std::move(response));
+        transport.send_request_to_server(::realm::internal::networking::to_request(std::move(request)),
+                                         [&p](auto&& response) {
+                                             p.set_value(::realm::internal::networking::to_core_response(std::move(response)));
                                          });
         return f.get();
     }